third_party/WebKit/LayoutTests/http/tests/devtools/cache-storage/cache-track-valid-origin.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/cache-storage/cache-storage-test.js"></script>
<script>

async function test() {
  var cacheStorageModel = InspectorTest.mainTarget.model(SDK.ServiceWorkerCacheModel);
  let invalidOrigins = ['http', 'test://fake', 'test://fake.origin.com', 'chrome://test'];
  let validOrigins = ['http://fake.origin.com', 'https://fake.origin.com'];

  InspectorTest.addResult('Invalid Origins:');
  invalidOrigins.map(origin => {
    InspectorTest.addResult(origin + ", valid = " + cacheStorageModel._isValidSecurityOrigin(origin));
  });
  InspectorTest.addResult('\nValid Origins:');
  validOrigins.map(origin => {
    InspectorTest.addResult(origin + ", valid = " + cacheStorageModel._isValidSecurityOrigin(origin));
  });
  InspectorTest.completeTest();
}
</script>
</head>
<body onload="runTest()">
<p>Tests that cache storage live update only tracks valid security origins.</p>
</body>
</html>
