third_party/WebKit/LayoutTests/http/tests/devtools/cache-storage/cache-track-valid-origin-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that cache storage live update only tracks valid security origins.

Invalid Origins:
http, valid = false
test://fake, valid = false
test://fake.origin.com, valid = false
chrome://test, valid = false

Valid Origins:
http://fake.origin.com, valid = true
https://fake.origin.com, valid = true
